requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
ffmpeg-python==0.2.0
filelock==3.10.0
future==0.18.3
idna==3.4
importlib-metadata==6.1.0
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
more-itertools==9.1.0
mpmath==1.3.0
networkx==3.0
numba==0.56.4
numpy==1.23.5
openai-whisper==20230314
PyAudio==0.2.13
regex==2022.10.31
requests==2.28.2
semantic-version==2.10.0
setuptools-rust==1.5.2
sympy==1.11.1
tiktoken==0.3.1
torch==2.0.0
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
zipp==3.15.0
